socios completo

+++ src/main/java/es/kokoro/controller/SociosController.java
@@ -1,11 +1,8 @@
 package es.kokoro.controller;
 
 import es.kokoro.dao.DAO;
-import es.kokoro.dao.mysql.MySQLSocioDAO;
 import es.kokoro.dao.mysql.MySqlFactoryDAO;
 import es.kokoro.model.Socio;
-import javafx.beans.property.Property;
-import javafx.beans.property.StringProperty;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
@@ -14,13 +11,12 @@ import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
-import javafx.scene.control.SortEvent;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.stage.Stage;
-
 import java.net.URL;
 import java.util.List;
 import java.util.ResourceBundle;
@@ -65,58 +61,71 @@ public class SociosController implements Initializable {
     @FXML
     public TableColumn Clmn_Estado;
 
-    private Socio socios;
-
-
-
     @FXML
     public void cerrarAccion(ActionEvent event) {
-
         Stage stage = (Stage) btn_cerrar.getScene().getWindow();
         stage.close();
-
     }
 
     @FXML
     public void modificarSocio(ActionEvent event2) throws Exception {
-        //TODO FALTA HACER
-        Stage stage2 = (Stage) btn_ModificarSocio.getScene().getWindow();
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("/view/ModificarSocio.fxml"));
-            Parent root1 = (Parent) fxmlLoader.load();
-            Stage stage = new Stage();
-            stage.setScene(new Scene(root1));
+        DAO<Socio> mySqlSocioDao = MySqlFactoryDAO.getMySqlFactoryDAO(Socio.class);
+        Socio socio = tbl_ListadoSocios.getSelectionModel().getSelectedItem();
 
-            stage.show();
-        } catch (Exception e) {
-            e.printStackTrace();
+        if (socio != null) {
+            Stage stage2 = (Stage) btn_ModificarSocio.getScene().getWindow();
+            try {
+                FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("/view/ModificarSocio.fxml"));
+                Parent root1 = (Parent) fxmlLoader.load();
+                ModificarSociosController modificarSociosController = fxmlLoader.getController();
+                modificarSociosController.setSocio(tbl_ListadoSocios.getSelectionModel().getSelectedItem());
+                Stage stage = new Stage();
+                stage.setScene(new Scene(root1));
+                stage2.close();
+                stage.show();
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+
+        } else {
+            Alert alert = new Alert(Alert.AlertType.WARNING);
+            alert.setTitle("Seleccion");
+            alert.setHeaderText("No ha seleccionado ningún socio");
+            alert.setContentText("Por favor, seleccione un socio de la tabla.");
+            alert.showAndWait();
         }
     }
 
     @FXML
     public void eliminarSocio(ActionEvent event3) throws Exception {
-
-        //TODO comprobar que hay una row seleccionada, falta sacar mensaje
-
         DAO<Socio> mySqlSocioDao = MySqlFactoryDAO.getMySqlFactoryDAO(Socio.class);
 
         Socio socioAEliminar = tbl_ListadoSocios.getSelectionModel().getSelectedItem();
 
-        mySqlSocioDao.delete(socioAEliminar);
-        tbl_ListadoSocios.getItems().remove(socioAEliminar);
+        if (socioAEliminar != null) {
+            mySqlSocioDao.delete(socioAEliminar);
+            tbl_ListadoSocios.getItems().remove(socioAEliminar);
 
+        } else {
+
+            Alert alert = new Alert(Alert.AlertType.WARNING);
+            alert.setTitle("Seleccion");
+            alert.setHeaderText("No ha seleccionado ningún socio");
+            alert.setContentText("Por favor, seleccione un socio de la tabla.");
+            alert.showAndWait();
+        }
     }
 
     @FXML
     public void nuevoSocio(ActionEvent event4) throws Exception {
-        //TODO FALTA HACER
+
         Stage stage4 = (Stage) btn_NuevoSocio.getScene().getWindow();
         try {
             FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("/view/AltaSocios.fxml"));
             Parent root1 = (Parent) fxmlLoader.load();
             Stage stage = new Stage();
             stage.setScene(new Scene(root1));
-
+            stage4.close();
             stage.show();
         } catch (Exception e) {
             e.printStackTrace();
@@ -138,17 +147,14 @@ public class SociosController implements Initializable {
         }
     }
 
-
     @Override
+    @FXML
     public void initialize(URL location, ResourceBundle resources) {
 
         try {
             DAO<Socio> mySQLSocioDAO = MySqlFactoryDAO.getMySqlFactoryDAO(Socio.class);
-
             List<Socio> socios = mySQLSocioDAO.getAll();
-
             ObservableList<Socio> sociosData = FXCollections.observableArrayList(socios);
-
             Clmn_idSocio.setCellValueFactory(new PropertyValueFactory<>("IdSocio"));
             Clmn_Nombre.setCellValueFactory(new PropertyValueFactory<>("Nombre"));
             Clmn_Apellidos.setCellValueFactory(new PropertyValueFactory<>("Apellidos"));
@@ -162,14 +168,11 @@ public class SociosController implements Initializable {
             Clmn_Periodo.setCellValueFactory(new PropertyValueFactory<>("Periodo"));
             Clmn_Cuota.setCellValueFactory(new PropertyValueFactory<>("Cuota"));
             Clmn_Estado.setCellValueFactory(new PropertyValueFactory<>("Estado"));
-
             tbl_ListadoSocios.setItems(sociosData);
 
         } catch (Exception e) {
             e.printStackTrace();
         }
-
-
 
     }
 }
